+++ src/bin/discover.rs
@@ -1,10 +1,8 @@
-use elgato_keylight::discover_elgato_devices;
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let devices = discover_elgato_devices().await?;
-
-    println!("{devices:#?}");
-
+    let devices = elgato_keylight::avahi::find_elgato_devices().await?;
+    for device in devices {
+        println!("{device}")
+    }
     Ok(())
 }